refactor auth structure
