+++ app/src/main/kotlin/com/stslex/plugins/auth/plugin/AdminPrincipal.kt
@@ -1,0 +1,8 @@
+package com.stslex.plugins.auth.plugin
+
+import io.ktor.server.auth.*
+
+data class AdminPrincipal(
+    val name: String,
+    val password: String
+) : Principal
